split ICU tokenizer into multiple files
