+++ src/nominatim_db/tokenizer/icu_tokenizer/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of Nominatim. (https://nominatim.org)
+#
+# Copyright (C) 2026 by the Nominatim developer community.
+# For a full list of authors see the git log.
+
+from .tokenizer import (create as create)
