+++ src/nominatim_db/tokenizer/icu_tokenizer/icu_types.py
