+++ test/python/tokenizer/test_icu_rule_loader.py
@@ -2,7 +2,7 @@
 #
 # This file is part of Nominatim. (https://nominatim.org)
 #
-# Copyright (C) 2025 by the Nominatim developer community.
+# Copyright (C) 2026 by the Nominatim developer community.
 # For a full list of authors see the git log.
 """
 Tests for converting a config file to ICU rules.
@@ -12,7 +12,7 @@ from textwrap import dedent
 import pytest
 import yaml
 
-from nominatim_db.tokenizer.icu_rule_loader import ICURuleLoader
+from nominatim_db.tokenizer.icu_tokenizer.rule_loader import ICURuleLoader
 from nominatim_db.errors import UsageError
 
 from icu import Transliterator
